+++ server/schedule/operator/kind.go
