+++ server/schedule/operator/step.go
@@ -1,0 +1,365 @@
+// Copyright 2019 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package operator
+
+import (
+	"bytes"
+	"fmt"
+
+	"github.com/pingcap/kvproto/pkg/metapb"
+	"github.com/pingcap/kvproto/pkg/pdpb"
+	"github.com/pingcap/log"
+	"github.com/pingcap/pd/server/core"
+	"go.uber.org/zap"
+)
+
+const (
+	// RegionInfluence represents the influence of a operator step, which is used by ratelimit.
+	RegionInfluence int64 = 1000
+	// smallRegionInfluence represents the influence of a operator step
+	// when the region size is smaller than smallRegionThreshold, which is used by ratelimit.
+	smallRegionInfluence int64 = 200
+	// smallRegionThreshold is used to represent a region which can be regarded as a small region once the size is small than it.
+	smallRegionThreshold int64 = 20
+)
+
+// OpStep describes the basic scheduling steps that can not be subdivided.
+type OpStep interface {
+	fmt.Stringer
+	ConfVerChanged(region *core.RegionInfo) bool
+	IsFinish(region *core.RegionInfo) bool
+	Influence(opInfluence OpInfluence, region *core.RegionInfo)
+}
+
+// TransferLeader is an OpStep that transfers a region's leader.
+type TransferLeader struct {
+	FromStore, ToStore uint64
+}
+
+// ConfVerChanged returns true if the conf version has been changed by this step
+func (tl TransferLeader) ConfVerChanged(region *core.RegionInfo) bool {
+	return false // transfer leader never change the conf version
+}
+
+func (tl TransferLeader) String() string {
+	return fmt.Sprintf("transfer leader from store %v to store %v", tl.FromStore, tl.ToStore)
+}
+
+// IsFinish checks if current step is finished.
+func (tl TransferLeader) IsFinish(region *core.RegionInfo) bool {
+	return region.GetLeader().GetStoreId() == tl.ToStore
+}
+
+// Influence calculates the store difference that current step makes.
+func (tl TransferLeader) Influence(opInfluence OpInfluence, region *core.RegionInfo) {
+	from := opInfluence.GetStoreInfluence(tl.FromStore)
+	to := opInfluence.GetStoreInfluence(tl.ToStore)
+
+	from.LeaderSize -= region.GetApproximateSize()
+	from.LeaderCount--
+	to.LeaderSize += region.GetApproximateSize()
+	to.LeaderCount++
+}
+
+// AddPeer is an OpStep that adds a region peer.
+type AddPeer struct {
+	ToStore, PeerID uint64
+}
+
+// ConfVerChanged returns true if the conf version has been changed by this step
+func (ap AddPeer) ConfVerChanged(region *core.RegionInfo) bool {
+	if p := region.GetStoreVoter(ap.ToStore); p != nil {
+		return p.GetId() == ap.PeerID
+	}
+	return false
+}
+func (ap AddPeer) String() string {
+	return fmt.Sprintf("add peer %v on store %v", ap.PeerID, ap.ToStore)
+}
+
+// IsFinish checks if current step is finished.
+func (ap AddPeer) IsFinish(region *core.RegionInfo) bool {
+	if p := region.GetStoreVoter(ap.ToStore); p != nil {
+		if p.GetId() != ap.PeerID {
+			log.Warn("obtain unexpected peer", zap.String("expect", ap.String()), zap.Uint64("obtain-voter", p.GetId()))
+			return false
+		}
+		return region.GetPendingVoter(p.GetId()) == nil
+	}
+	return false
+}
+
+// Influence calculates the store difference that current step makes.
+func (ap AddPeer) Influence(opInfluence OpInfluence, region *core.RegionInfo) {
+	to := opInfluence.GetStoreInfluence(ap.ToStore)
+
+	regionSize := region.GetApproximateSize()
+	to.RegionSize += regionSize
+	to.RegionCount++
+	if regionSize > smallRegionThreshold {
+		to.StepCost += RegionInfluence
+	} else if regionSize <= smallRegionThreshold && regionSize > core.EmptyRegionApproximateSize {
+		to.StepCost += smallRegionInfluence
+	}
+}
+
+// AddLearner is an OpStep that adds a region learner peer.
+type AddLearner struct {
+	ToStore, PeerID uint64
+}
+
+// ConfVerChanged returns true if the conf version has been changed by this step
+func (al AddLearner) ConfVerChanged(region *core.RegionInfo) bool {
+	if p := region.GetStorePeer(al.ToStore); p != nil {
+		return p.GetId() == al.PeerID
+	}
+	return false
+}
+
+func (al AddLearner) String() string {
+	return fmt.Sprintf("add learner peer %v on store %v", al.PeerID, al.ToStore)
+}
+
+// IsFinish checks if current step is finished.
+func (al AddLearner) IsFinish(region *core.RegionInfo) bool {
+	if p := region.GetStoreLearner(al.ToStore); p != nil {
+		if p.GetId() != al.PeerID {
+			log.Warn("obtain unexpected peer", zap.String("expect", al.String()), zap.Uint64("obtain-learner", p.GetId()))
+			return false
+		}
+		return region.GetPendingLearner(p.GetId()) == nil
+	}
+	return false
+}
+
+// Influence calculates the store difference that current step makes.
+func (al AddLearner) Influence(opInfluence OpInfluence, region *core.RegionInfo) {
+	to := opInfluence.GetStoreInfluence(al.ToStore)
+
+	regionSize := region.GetApproximateSize()
+	to.RegionSize += regionSize
+	to.RegionCount++
+	if regionSize > smallRegionThreshold {
+		to.StepCost += RegionInfluence
+	} else if regionSize <= smallRegionThreshold && regionSize > core.EmptyRegionApproximateSize {
+		to.StepCost += smallRegionInfluence
+	}
+}
+
+// PromoteLearner is an OpStep that promotes a region learner peer to normal voter.
+type PromoteLearner struct {
+	ToStore, PeerID uint64
+}
+
+// ConfVerChanged returns true if the conf version has been changed by this step
+func (pl PromoteLearner) ConfVerChanged(region *core.RegionInfo) bool {
+	if p := region.GetStoreVoter(pl.ToStore); p != nil {
+		return p.GetId() == pl.PeerID
+	}
+	return false
+}
+
+func (pl PromoteLearner) String() string {
+	return fmt.Sprintf("promote learner peer %v on store %v to voter", pl.PeerID, pl.ToStore)
+}
+
+// IsFinish checks if current step is finished.
+func (pl PromoteLearner) IsFinish(region *core.RegionInfo) bool {
+	if p := region.GetStoreVoter(pl.ToStore); p != nil {
+		if p.GetId() != pl.PeerID {
+			log.Warn("obtain unexpected peer", zap.String("expect", pl.String()), zap.Uint64("obtain-voter", p.GetId()))
+		}
+		return p.GetId() == pl.PeerID
+	}
+	return false
+}
+
+// Influence calculates the store difference that current step makes.
+func (pl PromoteLearner) Influence(opInfluence OpInfluence, region *core.RegionInfo) {}
+
+// RemovePeer is an OpStep that removes a region peer.
+type RemovePeer struct {
+	FromStore uint64
+}
+
+// ConfVerChanged returns true if the conf version has been changed by this step
+func (rp RemovePeer) ConfVerChanged(region *core.RegionInfo) bool {
+	return region.GetStorePeer(rp.FromStore) == nil
+}
+
+func (rp RemovePeer) String() string {
+	return fmt.Sprintf("remove peer on store %v", rp.FromStore)
+}
+
+// IsFinish checks if current step is finished.
+func (rp RemovePeer) IsFinish(region *core.RegionInfo) bool {
+	return region.GetStorePeer(rp.FromStore) == nil
+}
+
+// Influence calculates the store difference that current step makes.
+func (rp RemovePeer) Influence(opInfluence OpInfluence, region *core.RegionInfo) {
+	from := opInfluence.GetStoreInfluence(rp.FromStore)
+
+	from.RegionSize -= region.GetApproximateSize()
+	from.RegionCount--
+}
+
+// MergeRegion is an OpStep that merge two regions.
+type MergeRegion struct {
+	FromRegion *metapb.Region
+	ToRegion   *metapb.Region
+	// there are two regions involved in merge process,
+	// so to keep them from other scheduler,
+	// both of them should add MerRegion operatorStep.
+	// But actually, TiKV just needs the region want to be merged to get the merge request,
+	// thus use a IsPassive mark to indicate that
+	// this region doesn't need to send merge request to TiKV.
+	IsPassive bool
+}
+
+// ConfVerChanged returns true if the conf version has been changed by this step
+func (mr MergeRegion) ConfVerChanged(region *core.RegionInfo) bool {
+	return false
+}
+
+func (mr MergeRegion) String() string {
+	return fmt.Sprintf("merge region %v into region %v", mr.FromRegion.GetId(), mr.ToRegion.GetId())
+}
+
+// IsFinish checks if current step is finished.
+func (mr MergeRegion) IsFinish(region *core.RegionInfo) bool {
+	if mr.IsPassive {
+		return !bytes.Equal(region.GetStartKey(), mr.ToRegion.StartKey) || !bytes.Equal(region.GetEndKey(), mr.ToRegion.EndKey)
+	}
+	return false
+}
+
+// Influence calculates the store difference that current step makes.
+func (mr MergeRegion) Influence(opInfluence OpInfluence, region *core.RegionInfo) {
+	if mr.IsPassive {
+		for _, p := range region.GetPeers() {
+			o := opInfluence.GetStoreInfluence(p.GetStoreId())
+			o.RegionCount--
+			if region.GetLeader().GetId() == p.GetId() {
+				o.LeaderCount--
+			}
+		}
+	}
+}
+
+// SplitRegion is an OpStep that splits a region.
+type SplitRegion struct {
+	StartKey, EndKey []byte
+	Policy           pdpb.CheckPolicy
+	SplitKeys        [][]byte
+}
+
+// ConfVerChanged returns true if the conf version has been changed by this step
+func (sr SplitRegion) ConfVerChanged(region *core.RegionInfo) bool {
+	return false
+}
+
+func (sr SplitRegion) String() string {
+	return fmt.Sprintf("split region with policy %s", sr.Policy.String())
+}
+
+// IsFinish checks if current step is finished.
+func (sr SplitRegion) IsFinish(region *core.RegionInfo) bool {
+	return !bytes.Equal(region.GetStartKey(), sr.StartKey) || !bytes.Equal(region.GetEndKey(), sr.EndKey)
+}
+
+// Influence calculates the store difference that current step makes.
+func (sr SplitRegion) Influence(opInfluence OpInfluence, region *core.RegionInfo) {
+	for _, p := range region.GetPeers() {
+		inf := opInfluence.GetStoreInfluence(p.GetStoreId())
+		inf.RegionCount++
+		if region.GetLeader().GetId() == p.GetId() {
+			inf.LeaderCount++
+		}
+	}
+}
+
+// AddLightPeer is an OpStep that adds a region peer without considering the influence.
+type AddLightPeer struct {
+	ToStore, PeerID uint64
+}
+
+// ConfVerChanged returns true if the conf version has been changed by this step
+func (ap AddLightPeer) ConfVerChanged(region *core.RegionInfo) bool {
+	if p := region.GetStoreVoter(ap.ToStore); p != nil {
+		return p.GetId() == ap.PeerID
+	}
+	return false
+}
+
+func (ap AddLightPeer) String() string {
+	return fmt.Sprintf("add peer %v on store %v", ap.PeerID, ap.ToStore)
+}
+
+// IsFinish checks if current step is finished.
+func (ap AddLightPeer) IsFinish(region *core.RegionInfo) bool {
+	if p := region.GetStoreVoter(ap.ToStore); p != nil {
+		if p.GetId() != ap.PeerID {
+			log.Warn("obtain unexpected peer", zap.String("expect", ap.String()), zap.Uint64("obtain-voter", p.GetId()))
+			return false
+		}
+		return region.GetPendingVoter(p.GetId()) == nil
+	}
+	return false
+}
+
+// Influence calculates the store difference that current step makes.
+func (ap AddLightPeer) Influence(opInfluence OpInfluence, region *core.RegionInfo) {
+	to := opInfluence.GetStoreInfluence(ap.ToStore)
+
+	to.RegionSize += region.GetApproximateSize()
+	to.RegionCount++
+}
+
+// AddLightLearner is an OpStep that adds a region learner peer without considering the influence.
+type AddLightLearner struct {
+	ToStore, PeerID uint64
+}
+
+// ConfVerChanged returns true if the conf version has been changed by this step
+func (al AddLightLearner) ConfVerChanged(region *core.RegionInfo) bool {
+	if p := region.GetStorePeer(al.ToStore); p != nil {
+		return p.GetId() == al.PeerID
+	}
+	return false
+}
+
+func (al AddLightLearner) String() string {
+	return fmt.Sprintf("add learner peer %v on store %v", al.PeerID, al.ToStore)
+}
+
+// IsFinish checks if current step is finished.
+func (al AddLightLearner) IsFinish(region *core.RegionInfo) bool {
+	if p := region.GetStoreLearner(al.ToStore); p != nil {
+		if p.GetId() != al.PeerID {
+			log.Warn("obtain unexpected peer", zap.String("expect", al.String()), zap.Uint64("obtain-learner", p.GetId()))
+			return false
+		}
+		return region.GetPendingLearner(p.GetId()) == nil
+	}
+	return false
+}
+
+// Influence calculates the store difference that current step makes.
+func (al AddLightLearner) Influence(opInfluence OpInfluence, region *core.RegionInfo) {
+	to := opInfluence.GetStoreInfluence(al.ToStore)
+
+	to.RegionSize += region.GetApproximateSize()
+	to.RegionCount++
+}
